Correction des erreurs de compilation avec des drapeaux strictes de compilation.

+++ biblinternes/image/operations/champs_distance.h
@@ -297,7 +297,7 @@ void initialise(
 {
 	const auto sx = image.nombre_colonnes();
 	const auto sy = image.nombre_lignes();
-	const auto max_dist = dls::math::hypotenuse(static_cast<float>(sx), static_cast<float>(sy));
+	const auto max_dist = static_cast<int>(dls::math::hypotenuse(static_cast<float>(sx), static_cast<float>(sy)));
 
 	for (auto l = 0; l < sy; ++l) {
 		for (auto c = 0; c < sx; ++c) {
